+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Merged [Search-by-Distance-SMP](https://github.com/regorxxx/Search-by-Distance-SMP) script.
 - Merged [ListenBrainz-SMP](https://github.com/regorxxx/ListenBrainz-SMP) script.
 - Merged [Autobackup-SMP](https://github.com/regorxxx/Autobackup-SMP) script.
+- Merged [Device-Priority-SMP](https://github.com/regorxxx/Device-Priority-SMP) script.
 ### Changed
 ### Removed
 ### Fixed